task 0 - update

+++ 0x1E-search_algorithms/0-linear.c
@@ -11,16 +11,16 @@
 int linear_search(int *array, size_t size, int value)
 {
 
-	size_t index;
+	int index;
 
 	if (array == NULL)
 	{
 		return (-1);
 	}
 
-	for (index = 0; index < size; index++)
+	for (index = 0; index < (int)size; index++)
 	{
-		printf("Value checked array[%u] = [%d]\n", index, array[index]);
+		printf("Value checked array[%d] = [%d]\n", index, array[index]);
 
 		if (array[index] == value)
 		{
